Using cereal for serialization Battery curve linerization Fixed many warnings Refreshing the device info data after pairing as well Compiling with VS2019

+++ manager/src/ExportDataDialog.cpp
@@ -263,7 +263,7 @@ bool ExportDataDialog::exportTo(std::ostream& stream, size_t maxCount, bool show
     std::unique_ptr<QProgressDialog> progressDialog;
     if (showProgress)
     {
-        progressDialog.reset(new QProgressDialog("Exporting data...", "Abort", 0, size / 16, this));
+        progressDialog.reset(new QProgressDialog("Exporting data...", "Abort", 0, (int)size / 16, this));
         progressDialog->setWindowModality(Qt::WindowModal);
         progressDialog->setMinimumDuration(1000);
     }
@@ -272,7 +272,7 @@ bool ExportDataDialog::exportTo(std::ostream& stream, size_t maxCount, bool show
     {
         if (progressDialog && (i & 0xF) == 0)
         {
-            progressDialog->setValue(i / 16);
+            progressDialog->setValue((int)i / 16);
             if (progressDialog->wasCanceled())
             {
                 return false;
